カスタムフック作成、ファイル削除等
