+++ .env.production
@@ -1,0 +1,10 @@
+NEXT_PUBLIC_FIREBASE_API_KEY="*****"
+NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN="******"
+NEXT_PUBLIC_FIREBASE_PROJECT_ID="*****"
+NEXT_PUBLIC_FIREBASE_STORAGE_BUCKET="****"
+NEXT_PUBLIC_FIREBASE_MESSAGING_SENDER_ID="**********"
+NEXT_PUBLIC_FIREBASE_APP_ID="*****"
+NEXT_PUBLIC_FIREBASE_MEASUREMENT_ID="********"
+
+#ここからはlocalと異なる設定
+NEXT_PUBLIC_API_HOST = http://hogehoge.site
