--- frontend/src/components/DraggableTextBox.module.css
@@ -1,99 +1,0 @@
-.textBox {
-  position: relative;
-  border: 1px solid #ccc;
-}
-
-.textBox:hover .resizeHandles {
-  display: block;
-}
-
-.resizeHandles {
-  z-index: -1;
-  display: none;
-  position: absolute;
-  inset: 0;
-}
-
-.resizeHandleSouthEast {
-  position: absolute;
-  bottom: -5px;
-  right: -5px;
-  width: 10px;
-  height: 10px;
-  background-color: blue;
-  cursor: se-resize;
-}
-
-.resizeHandleSouth {
-  position: absolute;
-  bottom: -5px;
-  left: 50%;
-  width: 10px;
-  transform: translateX(-50%);
-  height: 10px;
-  background-color: blue;
-  cursor: s-resize;
-}
-
-.resizeHandleSouthWest {
-  position: absolute;
-  bottom: -5px;
-  left: -5px;
-  width: 10px;
-  height: 10px;
-  background-color: blue;
-  cursor: sw-resize;
-}
-
-.resizeHandleWest {
-  position: absolute;
-  top: 50%;
-  left: -5px;
-  width: 10px;
-  transform: translateY(-50%);
-  height: 10px;
-  background-color: blue;
-  cursor: w-resize;
-}
-
-.resizeHandleNorthWest {
-  position: absolute;
-  top: -5px;
-  left: -5px;
-  width: 10px;
-  height: 10px;
-  background-color: blue;
-  cursor: nw-resize;
-}
-
-.resizeHandleNorth {
-  position: absolute;
-  top: -5px;
-  left: 50%;
-  width: 10px;
-  transform: translateX(-50%);
-  height: 10px;
-  background-color: blue;
-  cursor: n-resize;
-}
-
-.resizeHandleNorthEast {
-  position: absolute;
-  top: -5px;
-  right: -5px;
-  width: 10px;
-  height: 10px;
-  background-color: blue;
-  cursor: ne-resize;
-}
-
-.resizeHandleEast {
-  position: absolute;
-  top: 50%;
-  right: -5px;
-  width: 10px;
-  transform: translateY(-50%);
-  height: 10px;
-  background-color: blue;
-  cursor: e-resize;
-}
