--- frontend/src/components/SlideEditor.module.css
@@ -1,17 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  background-color: #f2f2f2;
-  padding: 10px;
-  width: 100%;
-  height: 100vh;
-  border: 1px solid #ccc;
-}
-
-.editSpace {
-  border: 1px solid #ccc;
-  background-color: #fff;
-  width: 90%;
-  height: 90%;
-  position: relative;
-}
